move fmt_compat/userver_fmt_compat.hpp to shared/include/userver/utils/fmt_compat.hpp
