+++ shared/include/userver/utils/boost_uuid4.hpp
@@ -6,7 +6,7 @@
 #include <boost/uuid/uuid.hpp>
 
 #include <fmt/core.h>
-#include <userver_fmt_compat.hpp>
+#include <userver/utils/fmt_compat.hpp>
 
 USERVER_NAMESPACE_BEGIN
