+++ shared/include/userver/utils/strong_typedef.hpp
@@ -9,7 +9,7 @@
 #include <utility>
 
 #include <fmt/format.h>
-#include <userver_fmt_compat.hpp>
+#include <userver/utils/fmt_compat.hpp>
 
 #include <boost/functional/hash_fwd.hpp>
